python-rag-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
python-multipart==0.0.20
dashscope==1.23.2
pymilvus==2.5.6
beautifulsoup4==4.13.4
pypdf==5.4.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
httpx==0.28.1
pytest==8.3.5
